requirements.txt
gymnasium>=0.29.0
numpy>=1.26.0
pyyaml>=6.0
rich>=13.0.0
pytest>=8.0.0
